add '-build'

+++ scss/scss.js
@@ -1,12 +1,13 @@
 #! /usr/bin/env node
 
 /*
-1. 载入已写好的配置, 侦听多个项目: sudo /home/leon/share/tools/scss.js
-2. 针对临时项目(带include_paths): sudo /home/leon/share/tools/scss.js /home/leon//share/tools/test/p /home/leon//share/tools/test/p/f/
-当不会用到@import导入时, 第3个参数可以省略;
+1. sudo /path/scss.js # 载入已写好的配置, 侦听多个项目;
+   推荐这种方法, 把多个项目的配置信息写到 scssConfig.js中, 不需要额外的参数.
+2. sudo /path/scss.js /path/project1/ /path/lib/ # 临时监听project1项目, lib为.scss里@import的path(无@import,可以省略);
+3. sudo /path/scss.js -clear # 清除错误日志(如果配置了 scssConfig.js里logDir的路径);
+4. sudo /path/scss.js -build # 读取所有的.scss并编译成.css文件;
 
-推荐使用第1种, 只需把配置写到 scssConfig.js中, 不需要额外的参数.
-
+scss docs:
 http://sass-lang.com/docs/yardoc/file.SASS_REFERENCE.html
 */
 
@@ -24,7 +25,7 @@ node-sass使用时遇到的问题整理:
 var argv = process.argv,
     arguments = argv.splice(2),
     projectDir = arguments[0],
-    includePaths = arguments.length > 1 ? [arguments[1]] : []; // 一个文件夹应该够了
+    includePaths = arguments.length > 1 ? [arguments[1]] : []; // 单个项目文件夹
 
 // console.log(projectDir);
 
@@ -33,33 +34,39 @@ var $path = require('path');
 var $exec = require('child_process').exec;
 var $sass = require('node-sass');
 
-var moment = require('moment');
-// console.log(moment().format('YYYY/MM/DD HH:mm:ss'));
+var moment = require('moment'); // moment().format('YYYY/MM/DD HH:mm:ss')
+
+var isCommand = 0 === projectDir.indexOf('-') ? true : false;
 
 var scssConfig, logDir, logFile, maps;
-if (projectDir && '-clear' !== projectDir) { // one
-    watch2(projectDir, includePaths);
+
+if (projectDir && !isCommand) { // not: -clear || -build
+    watch(projectDir, includePaths);
 } else { // load maps
     scssConfig = require(__dirname + '/scssConfig');
     logDir = scssConfig.logDir;
     maps = scssConfig.maps;
     logFile = $path.join(logDir, 'scss.txt');
     // console.log(maps);
-    if ('-clear' === projectDir && logDir) { // clear logs
-        $fs.existsSync(logFile) && $fs.unlinkSync(logFile);
-        return;
+    if (isCommand) {
+        if ('-clear' === projectDir && logDir) { // clear logs
+            $fs.existsSync(logFile) && $fs.unlinkSync(logFile);
+        }
+        if ('-build' === projectDir) { // build .scss files
+            maps.forEach(function(item, i) {
+                var dir = item.dir,
+                    paths = item.paths;
+                build(dir, paths);
+            });
+        }
+    } else { // watch project
+        maps.forEach(function(item, i) {
+            var dir = item.dir,
+                paths = item.paths;
+            watch(dir, paths);
+        });
     }
-    maps.forEach(function(item, i) {
-        var dir = item.dir,
-            paths = item.paths;
-        watch2(dir, paths);
-    });
 }
-
-/*
-sass.render('body{background:blue; a{color:black;}}', function(err, css){
-    console.log(css);
-});*/
 
 var AP = Array.prototype;
 var logMsg = function() {
@@ -69,7 +76,7 @@ var logMsg = function() {
         isErr = true === last ? true : false;
     if (isErr) args.pop();
     if (isErr && logDir) {
-        var data = '[' + moment().format('YYYY/MM/DD HH:mm:ss') + ']'+ '\n' + AP.join.call(args, '\n');
+        var data = '[' + moment().format('YYYY/MM/DD HH:mm:ss') + ']' + '\n' + AP.join.call(args, '\n');
         data = data.replace(/\n$/g, '');
         $fs.appendFileSync(logFile, data + '\n', 'utf-8');
     }
@@ -81,12 +88,12 @@ var processFile = function(filepath, paths) {
     var baseName = $path.basename(filepath, '.scss'),
         target = $path.join($path.dirname(filepath), baseName + '.css');
     // not *.scss and _*.scss files
-    if ('.scss' !== $path.extname(filepath) || baseName.indexOf('_') === 0) return;
+    if ('.scss' !== $path.extname(filepath) || 0 === baseName.indexOf('_')) return;
 
     // logMsg(filepath, ' > ', target);
     // return;
     var saveFile = function(data) {
-        $fs.writeFile(target, data, function (err) {
+        $fs.writeFile(target, data, function(err) {
             if (err) return logMsg(err, true);
             // logMsg('data', data);
             logMsg('@done:', filepath, '>', target);
@@ -103,46 +110,10 @@ var processFile = function(filepath, paths) {
     });
 };
 
-function watch(file, real) {
-    // watch folder && subfolders with fs.watch: https://github.com/remy/nodemon/issues/66
-    // or node-walk: https://github.com/coolaj86/node-walk
-    var path;
-    if (!real) {
-        if (!$fs.existsSync(file)) return logMsg('No such directory: ' + file, true);
-        real = $fs.realpathSync(file);
-        path = real;
-    } else {
-        path = $path.join(real, file);//real.concat('/', file);
-    }
-    real = path;
-    $fs.stat(path, function(err, stat) {
-        if (err) return logMsg(err, true);
-        if (stat && stat.isDirectory()) {
-            $fs.watch(path, function(event, filename) {
-                var filepath = $path.join(real, filename);
-                // var type  = event == 'change' ? 'changed' : 'created';
-                // logMsg(event, filename, filepath);
-                if (filename) processFile(filepath);
-            });
-            // sub folders
-            $fs.readdir(path, function(err, files) {
-                if (err) return logMsg(err, true);
-                for (var i = 0, len = files.length; i < len; i++) {
-                    var file = files[i],
-                        ext = $path.extname(file);
-                    // logMsg(file, real);
-                    if (!ext) { // folder
-                        watch(file, real);
-                    }
-                }
-            });
-        }
-    });
-};
-function watch2(path, paths) {
+function watch(path, paths) {
     $exec('find ' + path + ' -type d', function(err, stdout, stderr) {
         if (err) return logMsg(err, true);
-        var dirs =  stdout.split('\n');
+        var dirs = stdout.split('\n');
         // logMsg(dirs);
         dirs.forEach(function(dirpath, i) {
             // logMsg(i, dirpath);
@@ -154,4 +125,16 @@ function watch2(path, paths) {
             }
         });
     });
-};
+}
+
+function build(path, paths) {
+    $exec('find ' + path + ' -name "*.scss"', function(err, stdout, stderr) {
+        if (err) return logMsg(err, true);
+        var files = stdout.split('\n');
+        // logMsg(files);
+        files.forEach(function(filepath, i) {
+            // logMsg(i, filepath);
+            if (filepath) processFile(filepath, paths);
+        });
+    });
+}
